Переписал все на mybatis

+++ src/main/resources/schema.sql
@@ -16,15 +16,15 @@ create table if not exists tasks(
 );
 
 create table if not exists users_tasks(
-    user_id int not null,
-    task_id int not null,
+    user_id bigint not null,
+    task_id bigint not null,
     primary key(user_id, task_id),
     constraint fk_users_tasks_users foreign key(user_id) references users(id) on delete cascade on update no action,
     constraint fk_users_tasks_tasks foreign key(task_id) references tasks(id) on delete cascade on update no action
 );
 
 create table if not exists users_roles(
-    user_id int not null,
+    user_id bigint not null,
     role varchar(100) not null,
     primary key(user_id, role),
     constraint fk_users_roles_users foreign key(user_id) references users(id) on delete cascade on update no action
